+++ examples/example_both.rs
@@ -1,5 +1,4 @@
 use std::{
-    net::{IpAddr, Ipv4Addr},
     sync::{
         atomic::{AtomicBool, Ordering},
         Arc,
@@ -32,7 +31,7 @@ enum MessageServerToClient {
 fn client() {
     let server_address = "127.0.0.1:21001".parse().unwrap();
     // Full type is `ClientNetworking<MessageClientToServer, MessageServerToClient>`.
-    let mut client = ClientNetworking::new(server_address);
+    let mut client = ClientNetworking::new(server_address, None);
 
     let mut last_sent_time = Instant::now();
 
@@ -98,9 +97,8 @@ fn client() {
 
 fn server() {
     let desired_port = 21001;
-    let address = IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1));
     // Full type is `ServerListenerNetworking<MessageServerToClient, MessageClientToServer>`.
-    let server_listener = ServerListenerNetworking::new(address, desired_port);
+    let server_listener = ServerListenerNetworking::new(Some(desired_port), None, None);
     let actual_port = server_listener.server_port();
     println!("Opened on port {actual_port}");