+++ examples/example_server.rs
@@ -1,5 +1,4 @@
 use std::{
-    net::{IpAddr, Ipv4Addr},
     sync::{
         atomic::{AtomicBool, Ordering},
         Arc,
@@ -26,9 +25,8 @@ enum MessageServerToClient {
 
 fn main() {
     let desired_port = 21001;
-    let address = IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1));
     // Full type is `ServerListenerNetworking<MessageServerToClient, MessageClientToServer>`.
-    let server_listener = ServerListenerNetworking::new(address, desired_port);
+    let server_listener = ServerListenerNetworking::new(Some(desired_port), None, None);
     let actual_port = server_listener.server_port();
     println!("Opened on port {actual_port}");